+++ src/id_translation/fetching/types.py
@@ -1,29 +1,36 @@
 """Types related to translation fetching."""
 
-from dataclasses import dataclass
-from typing import Generic, Iterable, Optional, Set
+from dataclasses import dataclass as _dataclass
+from typing import Generic as _Generic, Iterable, Optional, Set
 
 from ..offline.types import PlaceholdersTuple
 from ..types import IdType, SourceType
 
 
-@dataclass(frozen=True)
-class IdsToFetch(Generic[SourceType, IdType]):
+@_dataclass(frozen=True)
+class IdsToFetch(_Generic[SourceType, IdType]):
     """A source and the IDs to fetch from it."""
 
     source: SourceType
     """Where to fetch from."""
-    ids: Optional[Iterable[IdType]]
-    """Unique IDs to fetch translations for. Fetch as much as possible if ``None``"""
+    ids: Iterable[IdType]
+    """Unique IDs to fetch translations for.``"""
 
 
-@dataclass(frozen=True)
-class FetchInstruction(IdsToFetch[SourceType, IdType]):
+@_dataclass(frozen=True)
+class FetchInstruction(_Generic[SourceType, IdType]):
     """Instructions passed from an ``AbstractFetcher`` to an implementation."""
 
+    source: SourceType
+    """Where to fetch from."""
     placeholders: PlaceholdersTuple
     """All desired placeholders in preferred order."""
     required: Set[str]
     """Placeholders that must be included in the response."""
-    all_placeholders: bool
-    """Flag indicated whether to retrieve as many placeholders as possible."""
+    ids: Optional[Iterable[IdType]]
+    """Unique IDs to fetch translations for. Fetch as much as possible if ``None``."""
+
+    @property
+    def fetch_all(self) -> bool:
+        """Flag indicated that as much data as possible should be retrieved."""
+        return self.ids is None