+++ tests/conftest.py
@@ -1,7 +1,9 @@
+from functools import partialmethod
 from pathlib import Path
 from typing import Any, Dict, Iterable, List, Optional, Tuple
 
 import pytest
+from rics.mapping import Mapper
 
 from id_translation import Translator
 from id_translation.fetching import AbstractFetcher, support
@@ -11,6 +13,8 @@ from id_translation.offline import TranslationMap
 from id_translation.offline.types import PlaceholderTranslations
 
 ROOT: Path = Path(__file__).parent
+
+Mapper.__init__ = partialmethod(Mapper.__init__, verbose_logging=True)  # type: ignore[assignment]
 
 
 class HexFetcher(AbstractFetcher[str, int]):