+++ tests/fetching/test_abstract_fetcher.py
@@ -11,7 +11,7 @@ def fetcher(data):
 
 def test_unknown_sources(fetcher):
     with pytest.raises(exceptions.UnknownSourceError) as ec:
-        fetcher.fetch([IdsToFetch("humans", None), IdsToFetch("edible_humans", [1, 2])])
+        fetcher.fetch([IdsToFetch("humans", [1, 2]), IdsToFetch("edible_humans", [1, 2])])
     assert str({"edible_humans"}) in str(ec.value)
 
     with pytest.raises(exceptions.UnknownSourceError) as ec:
@@ -23,16 +23,10 @@ def test_fetch_all_forbidden(data):
     fetcher: AbstractFetcher[str, int] = MemoryFetcher(data)
     fetcher._allow_fetch_all = False
 
-    with pytest.raises(exceptions.ForbiddenOperationError) as ec:
-        fetcher.fetch([IdsToFetch("humans", None), IdsToFetch("animals", [1, 2])])
-    assert f"{repr(AbstractFetcher._FETCH_ALL)} not supported" in str(ec.value)
-
-    with pytest.raises(exceptions.ForbiddenOperationError) as ec:
+    with pytest.raises(exceptions.ForbiddenOperationError, match=f"{AbstractFetcher._FETCH_ALL!r} not supported"):
         fetcher.fetch_all()
-    assert f"{repr(AbstractFetcher._FETCH_ALL)} not supported" in str(ec.value)
 
 
 def test_unknown_placeholders(fetcher):
-    with pytest.raises(exceptions.UnknownPlaceholderError) as ec:
-        fetcher.fetch([IdsToFetch("humans", None)], ("id", "number_of_legs"), {"number_of_legs"})
-    assert "{'number_of_legs'} not recognized" in str(ec.value)
+    with pytest.raises(exceptions.UnknownPlaceholderError, match="{'number_of_legs'} not recognized"):
+        fetcher.fetch([IdsToFetch("humans", [])], ("id", "number_of_legs"), {"number_of_legs"})